Add logout button component and add useLayoutEffectmto set options in header
